Update patch set 10
Patch Set 10: Code-Review+1

I looked it over and it looks fine, the testing seems light, but this is also a light lib (not too much code).  I'm not sure where the python code is ci tested though.  Another potential issue (but may be fine as this is a small lib) is that the last commit was in november.  The repo itself was updated 4 days ago though (go code).

Patch-set: 10
Label: Code-Review=+1