Add upper constraint for setuptool >=39 on stable/newton
OpenStack stable/newton and older is now incompatible with the latest
version of setuptools so we need to add an upper constraint to prevent
us installing a version that is too new. Setuptools version 39.0.0
removes the ability to index the result of parse_version which is used
by oslo_utils.versionutils and therefore breaks anyone using those
utility functions.

Change-Id: I93e940538a943fba2d426a714a187e7f6390ceab