Update microversion-parse to 0.2.1
Early versions of microversion-parse were focused solely on parsing the
microversion header to extract a microversion for the current request.
Since then it has been used in association with some fairly generic
code to drive the microversion handling of the placement service. That
code is sufficiently genric that it has been moved into microversion-
parse so other people can use it.

Backwards compatibility has been maintain on the methods that
microversion-parse has exposed since the early days.

The updated version adds a dependency on WebOb, which is already in
global requirements. The minimum requirement in microversion-parse
is lower than the the minimum in global-requirements.txt

The placement service will be using features in the very latest
version of microversion parse, so the min is set here to the last
in both requirements and constraints.

microversion-parse remains apache 2.0 license, is managed in openstack
infra, has support for both python 2 and 3, and creates universal wheels.

Change-Id: I1b7d9357eee8f56f97b2dc8a894d360bf7cfc1ac