make the --reqs arg to project-requirements-change.py useful
The --reqs argument was not being used which made it difficult to run
the script locally to test it.

This change passes the directory given to --reqs as the requirements
repository directory so that in local mode the repository can live
anywhere.

The default is still set to the location used on the CI nodes.

Change-Id: Id812628345313b781d65140bb3c25adf1eb00bc5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>