Update patch set 2
Patch Set 2:

> Do we have any tests for this?

As far as I can tell there are no tests for this module at all. I did test locally using a patch to shade (https://review.openstack.org/555299) which was failing before and passes now. As part of testing the change to treatment of caps I added a cap to sphinx, verified that it was not allowed, and removed it.

Patch-set: 2