uncap pysaml2
This is the first step in updating pysaml2 to 4.5.0.
Only keystone needs to consume this before we update UC/GR (as they get a
pkg_resources.ContextualVersionConflict if we update everything in one go).

The versions were masked because they could bring in pycryptodome, which we do
not want.

Next step, after keystone consumes this change, is to update upper-constraints.

Partial-Bug: #1750843
Change-Id: I6cf9812deecbfaea5ff6e47ba3d48dd94db773b5