Update patch set 3
Patch Set 3: Code-Review+1

> > Do we still need to put doc requirements into the list, now that
 > we
 > > have a way for doc jobs to have their own set of dependencies?
 > 
 > If it's only used on doc/requirements.txt then I agree I don't
 > think we need this.

In the course of doing the lower-constraints work I found that doc/requirements.txt is still being checked against the g-r list. So I think we are going to need this after all, unless we want to change that.

Only +1 because the validation job is failing.

Patch-set: 3
Label: Code-Review=+1