Update patch set 2
Patch Set 2: Code-Review-1

I'd like to see a plan before we approve this.

Specifically how are we going to introduce newer versions of pysaml2 with paramiko etc.

In itself this doesn't enough for the check pipeline to tell us anything.  You need to bump u-c as well. to really see what happens to keystone, and all projects will need to work with old and new pysaml2s

Patch-set: 2
Label: Code-Review=-1