Update patch set 2
Patch Set 2:

> tony, the first patchset uncapped, raised uc and raised gr.  that
 > was depended on https://review.openstack.org/557440/ which passed. 
 > Keystone is the only project depending on pysaml2 as well (looking
 > at http://codesearch.openstack.org/?q=pysaml2&i=nope&files=.*requirements.*&repos=
 > ).

Yes only keystone uses pysaml2, but the reason we blocked 4.0.3 was because *it* installed pycrytodome which as you know installs into the same namespace as pycrypto meaning that we now have non-deterministic behaviour, constraints saves us somewhat.

We know all versions between 4.0.2 and 4.5.0 are bad so we need to make sure we bypass them.

Ideally wed block them in g-r but we can rely on constraints so we can just remove the cap in g-r and set u-c ===4.5.0 then if things go pear shaped we have a single change to revert. and then in say 2 weeks bump in the minimum here and in keystone.

In you commit message you imply we have to remove the cap in one change before we can bump u-c if that's true I'd like to know why

The plan is much more straight forward given that pysaml2 has made the switch to cryptography but nothing in this review explains *any* of that.

Patch-set: 2