Update patch set 2
Patch Set 2: Code-Review+2

There are some commit message Q & A we usually want that are detailed here:

https://git.openstack.org/cgit/openstack/requirements/tree/README.rst#n265

But since this was discussed on the ML and has a few of us involved, I'm OK with this as is. Might want to add that if you need to update for some reason. Or if someone else from the requirements team feels more strongly about that.

But looks good to me, and I would like to see this get added so we can move forward with cleaning up some of the doc jobs that are relying on module API documentation.

Patch-set: 2
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2