Update patch set 5
Patch Set 5:

There is a valid time that'd happen (mainly on stable branckes).

Sometimes we have the following releases available.
1.0; 1.1; 1.2; 1.3
The GR is set to >=1.0;!=1.0
This would make the effective minimum 1.1.  This happens most often on stable branches where we don't bump minimums but only exclude known bad releases.

Patch-set: 5