Update patch set 5
Patch Set 5:

> There is a valid time that'd happen (mainly on stable branckes).
 > 
 > Sometimes we have the following releases available.
 > 1.0; 1.1; 1.2; 1.3
 > The GR is set to >=1.0;!=1.0
 > This would make the effective minimum 1.1.  This happens most often
 > on stable branches where we don't bump minimums but only exclude
 > known bad releases.

The global minimum is not used for testing local lower constraints.

The *exclusions* from the local requirement specification are compared to the global requirement specification to ensure that the local values are a subset of the global values. The minimum is not included in that set, and those rules are not applied to the local lower constraint.

The local lower constraint is checked against the *local* requirement specification to ensure that those versions match exactly. That implies we always use ">=" to set the minimum, and that we don't do something like in your example where we spell ">" using 2 separate (conflicting) specifiers.

Patch-set: 5