remove lower bounds from global requirements
Now that we allow projects to diverge in their lower bounds, it no
longer makes sense for us to track "global" lower bounds values. We
need to retain the exclusion specifiers (using the "!=" operator) so
that the automatic constraint update calculation continues to produce
a set of values that we expect to work.

In addition to updating the data, this patch updates the validation
rules so that if a lower bound is included an error is reported.

Change-Id: If9da4636755fc089b2ec5f075e03767f84c34a67
Signed-off-by: Doug Hellmann <doug@doughellmann.com>