Update patch set 5
Patch Set 5:

> after discussing this in our meeting we would like the following.
 > 
 > As this is a rocky release that you are wishing to use we have to
 > reject it.
 > Please backport the change needed to the queens branch and release
 > a 3.25.1 version with the fix as a queens release.
 > Once that's done, make a constraints update.
 > Since this is a logging vuln that this fixes we do not wish to bump
 > the effective minimum.

The backport to oslo.concurrency is done (and linked in the commit message).

The release will cause the constraint update to be proposed.

The backport in cinder will need to work whether the version of oslo.concurrency has the new argument or not, so you'll need to take some care with that.

Patch-set: 5
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0