Add psycopg2-binary to global-requirements
psycopg2 2.8 will rename its wheel package to psycopg2-binary [1]. We
want to continue using this as it is the only way to avoid compiling
psycopg2 as a static library or installing the 'libpq' package.

* Is the library actively maintained?

Yes. This package is simply the binary package for psycopg2.

* Is the library good code?

Yes. It is the most popular PostgreSQL adapter for the Python
programming language.

* Is the library python 3 compatible?

Yes.

* Is the library license compatible?

Yes.

* Is the library already packaged in the distros we target?

Yes (via the standard 'psycopg2' package - this package is only
necessary if installing from PyPi)

* Is the function of this library already, et cetera?

The psycopg2 library can be used instead, but doing so will require
compiling this library as a static library or installing the 'libpq'
package.

[1] http://initd.org/psycopg/docs/install.html#binary-install-from-pypi

Change-Id: I02490605b82abcedae71ed9cc6796bfb866cddcf
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>