Update patch set 1
Patch Set 1: Code-Review+1

ok for me, a bugreference would be good to track potential removal of this masking. 

is this known to be a bug in libvirt-python and did upstream acknowledge that its going to be fixed in 4.2.1 ? trying to understand if this is stopgap or permanent.

Patch-set: 1
Label: Code-Review=+1