Update patch set 1
Patch Set 1: Code-Review+2

This is a silly situation to be in but each vendor/distributor will need to handle this there own way.

As Stephen says from a distro POV the package doesn't change so this should mostly be a no-op there.  We will have a bunch of clean-up to do in anything that uses psycopg2 before we can accept the 2.8.0 version into g-r.

Probably needs a list discussion, or at least a heads up

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2