Update patch set 1
Patch Set 1:

so you are increasing the pike capped version of tempest which is totally fine as tempest is branchless and all version should work with supported stable branches. 

But dependent pckg has to be upper consternation according to new tempest version.   

For current error, you have to update upper constraint for os-testr also to os-testr===1.0.0 that is also fine as os-testr also branchless.

Patch-set: 1
Label: Verified=0