Bump keyring to 12.1.0 in order to cap secretstorage
Recently a new secretstorage was released, which was requiring
python 3. this dependency was pulled in via keyring. keyring 12.1
was relasing with those dependencies corrected:

                   'secretstorage; sys_platform=="linux" and '
                   'python_version>="3.5"',
                   'secretstorage (<3); (sys_platform=="linux2" or '
                   'sys_platform=="linux") and python_version<"3.5"'

This excludes the incompatible versions for python2 and requires
the newer release for python3. Blacklist the older 12.0.0 release
to avoid somebody else hitting the same issue.

Change-Id: I98fdb1af7fd85aa9b9645b33916bdfbfe500c882