Add tap-as-a-service
Now it's published on PyPI.
Add it to global-requirements, following the similar change. [1]

This is necessary for networking-midonet.  See [2].
Note: this is not a library. It provides a Neutron service plugin
as some of Neutron subprojects like neutron-vpnaas do.
networking-midonet has a driver for this service plugin and
associated unit tests, which currently requires this.

This commit doesn't add it to blacklist.txt because:
	- as far as I know, networking-midonet is the only project which
	  uses tox_install to install tap-as-a-service
	- and its gate is currently broken and we plan to retire
	  tox_install to fix it. [2]

[1] https://review.openstack.org/#/c/552865/
[2] https://review.openstack.org/#/c/555788/

The following answers the questions on:
https://docs.openstack.org/project-team-guide/dependency-management.html#for-new-requirements

> Is the library actively maintained?

It's maintained.  Not very active these days though.

> Is the library good code?

It's following Neutron standard. (loosely)

> Is the library python 3 compatible?

Yes.

> Is the library license compatible?

Yes. Same license as Neutron.

> Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

https://packages.ubuntu.com/search?keywords=python-neutron-taas

I have no idea about Fedora.

> Is the function of this library already covered by other libraries in global-requirements.txt?

No.

> Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes. networking-midonet has a driver for this project.

> If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?

No. It isn't a library.
Ideally it should be cycle-with-milestones.

> Do I need to update anything else?

I'm not sure what this question means.

Change-Id: Ifed26d4cd9477b7d421718a95d475d85a9720e1f