Update patch set 2
Patch Set 2:

Since this is not an official project, releases for these are done by the individual project.

It could be proposed to come under governance and start using the release management team for processing releases, but that would be up to the team if they want to do and I don't believe is necessary for adding to requirements.

Patch-set: 2
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0