Update patch set 2
Patch Set 2:

This looks OK except that it adds the project to the blacklist which means any two projects that want to depend on this won't be ensured to have compatible versions in their requirements files. Is that what we want? I see that we do the same for other networking projects, and I'm not sure why we do that.

Patch-set: 2
Label: Verified=0