Update patch set 4
Patch Set 4: Code-Review-1

Should you be able to install all plugins at the same time?  If so, the plugins should not be listed in blacklist.txt.  If not they should stay.
If you CAN use tap-as-a-service outside of the plugin (which you indicated is true), you should not list it in blacklist.txt so we can ensure it is co-installable with the rest of openstack (we don't care that it's used in a plugin for the test).

Patch-set: 4