Allow Pygments in openstack/requirements
This change will help doc8 check rst files without error D000.

Is the library actively maintained?
Yes - https://bitbucket.org/birkenfeld/pygments-main

Is the library good code?
Yes - It is the defacto syntax highlighting code in Python

Is the library python 3 compatible?
Yes

Is the library license compatible?
Yes - BSD

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
Yes

Is the function of this library already covered by other libraries in global-requirements.txt?
No

Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes - for doc8 checking where there is code snippets.
It is also a transient dependancy of sphinx - https://github.com/sphinx-doc/sphinx/blob/master/setup.py#L21

Change-Id: I643c7632a02c87820c9c51f03fa34f0764384783
Close-bug: #1771479