Update patch set 2
Patch Set 2:

ok, sounds like you need it, but I'm still not sure why.  What in particular is pulling pygments in?  (where is the import)?  Also, if you wish to add a new library to global-requirements.txt you'll need to answer the questions from the following link in the commit message.

https://docs.openstack.org/project-team-guide/dependency-management.html#for-new-requirements

Patch-set: 2