Update patch set 3
Patch Set 3:

So The bottom line here, as I understand it, is that doc8 on the designate docs is currently ignoring all D000 errors and would like to stop ignoring them.

D000 is done via docutils/setuptools but it's an *optional* check with a hard error [1]

So for any project that wants to perform similar checks will also need Pygments available.

I don't think Pymnets qualifies as a linter so we'll need to add this properly and include it in {lower,upper}-constraints.txt

[1] http://repo.or.cz/docutils.git/blob/HEAD:/sandbox/gitpull/setuptools/docutils/utils/code_analyzer.py#l11

Patch-set: 3
Label: Verified=0