Adding ansible-runner to global-requirements
networking-ansible ML2 driver invokes ansible via an ASL 2.0 python
wrapper named ansible-runner.

Is the library actively maintained?
Yes - https://github.com/ansible/ansible-runner
I've been working with the maintainers and they've been very responsive

Is the library good code?
Yes - Documentation has been good and contributing has been easy
stability has been good and good coding practices appear to be
practiced.

Is the library python 3 compatible?
Yes
https://github.com/ansible/ansible-runner/commit/b86c50978e70526e952a1a19bc94997f99b98992

Is the library license compatible?
Yes - ASL 2.0

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
Fedora - Yes
Ubuntu - No

Is the function of this library already covered by other libraries in global-requirements.txt?
I don't believe so. Afaik it's the only distibuted lib that provides a
programatic interface to ansible.

Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes - networking-ansible invokes it directly to manage switching
infrastructure via Ansible Networking

Reason for this backport:
We want to support networking-ansible in the Queens release. ansible-runner
is used only by networking-ansible so we won't affect other projects with
this patch. With ansible-runner not being present in Queens we won't be
able to make networking-ansible working in the Queens release.

pexpect library also had to be bumped because ansible-runner has a
requirement on it given as >= 4.5

Conflicts:
	global-requirements.txt
	lower-constraints.txt
	upper-constraints.txt

Change-Id: I33725dbb61f0e6423ef358602822e2e6bcef4abd
(cherry picked from commit a449abac5d3336db765006d8cc6829fd4785dccb)