Update constraint for python-novaclient to 10.1.0
This constraint update was originally tried [1] awhile back and failed
because novaclient had removed some functionality that openstackclient
needed for providing floating ip and host commands. So we had to leave
the constraint at 9.1.1 pike for novaclient, but this prevented users
from being able to use the volume multi-attach feature out-of-the-box
in queens.

To improve the situation for users, the openstackclient stable branch
team allowed us to backport changes to stable/queens that would make
queens openstackclient work with novaclient 10.1.0 (queens). We have
requested a new release with those changes as a dependency here.

Once openstackclient 3.14.2 for queens is released, we would like to
update the constraint for novaclient queens to 10.1.0 to enable use
of the volume multi-attach feature out-of-the-box for queens.

[1] https://review.openstack.org/538070

Depends-On: https://review.openstack.org/571537

Change-Id: I6e6b5dd95b0d5f54b3e8e2fce5bf303e91de42b7