Update patch set 1
Patch Set 1:

> recheck

I'll ask you again to refrain from leaving -1 reviews without a reason for the negative review.  OpenStack tries to be a welcoming community, leaving a -1 is the same as saying "this change needs work".  If you do not provide details on what you see as being wrong with the change the owner can not attempt to address the issues.

If you have any questions on how to provide meaningful reviews please feel free to reach out to myself of the TC

Patch-set: 1