Update patch set 3
Patch Set 3:

(2 comments)

> Abandoned

... even so trying to answer the question :)

 > I am just unsure how we're meant to reconcile all the lower
 > constraints across all projects... but ... eh?

We no longer enforce all projects to agree on a minimum.  If some project needs 0.5.1 (and works with all newer versions) and another needs 0.5.2 then that's actually fine.  We rely on all projects agreeing on a maximum version (which is indicated via upper-constraints.txt).

So for olso.serialization you can merge this and then something like:

 $ git diff
 diff --git a/lower-constraints.txt b/lower-constraints.txt
 index 4791fab..d2a212f 100644
 --- a/lower-constraints.txt
 +++ b/lower-constraints.txt
 @@ -17,7 +17,7 @@ mccabe==0.2.1
  mock==2.0.0
  monotonic==0.6
  mox3==0.20.0
 -msgpack==0.5.1
 +msgpack==0.5.2
  netaddr==0.7.18
  netifaces==0.10.4
  os-client-config==1.28.0
 diff --git a/requirements.txt b/requirements.txt
 index 25e2a76..abdd118 100644
 --- a/requirements.txt
 +++ b/requirements.txt
 @@ -9,6 +9,6 @@
  
  pbr!=2.1.0,>=2.0.0 # Apache-2.0
  six>=1.10.0 # MIT
 -msgpack>=0.5.1 # Apache-2.0
 +msgpack>=0.5.2 # Apache-2.0
  oslo.utils>=3.33.0 # Apache-2.0
  pytz>=2013.6 # MIT

is oslo.serialization. and all should be okay.

Patch-set: 3