Adding faker to global requirements
fixtures-git uses faker to generate fake text for inserting in to files
when building test git repos for tools that wish to test with git.

Is the library actively maintained?
Yes - https://github.com/joke2k/faker
They appear to have made regular releases since 2016 averaging ~2 per
month.

Is the library good code?
Yes - Appears to be relatively straight forward to extend suggesting a
nice simple design and the documentation appears to cover most of the
different types of fake date in reasonable detail.

Is the library python 3 compatible?
Yes - though only for >=3.4
https://github.com/joke2k/faker/blob/da96379c07dbd22bea070db5a8a3e1ad85e4cf04/setup.py#L43

Is the library license compatible?
Yes - MIT

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?

Fedora - Present since Fedora 27
Ubuntu - Present since 17.10

Is the function of this library already covered by other libraries in
global-requirements.txt?

I don't believe so, was unable to locate any other random text
generators.

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes - fixtures-git uses it directly to generate text to be used for
testing by developer tools working with git repositories such as
git-upstream.

Change-Id: Ic1c0412fb1473465977af8244b3e61060296e7a1