Add py36 testenv
Python 3.6 is installed by default in Ubuntu 18.04 LTS.
According to Python2 Deprecation Timeline [1] and Python3-first Goal [2],
we should have py36 testenv ready for it.

This change adds a non-voting job for checking the 3.6
upper-constraints, and a voting job for py36.

Once we're confident we have all the bits in place for 36-uc checking we
can promote that job to voting.

[1] https://governance.openstack.org/tc/resolutions/20180529-python2-deprecation-timeline.html
[2] https://review.openstack.org/#/c/575933

Change-Id: Ia06c1ab217e80902adb607a3bc6063aae66f48bf