Update patch set 1
Patch Set 1:

I'm not certain we want this.  I think a better approach is to just have a py3-XX target and set the basepaython to python3 and then let the distro provide a version.  Then when we run on trusty/xenial/bionic we get 3.4/3.5/3.6 and don't need to track it ourselves.

I know this has come up before.  We probably need to circle back to os-dev and the python-first community goal before approving this

Patch-set: 1
Label: Verified=0