Update patch set 3
Patch Set 3: Code-Review+1

I did a little digging and the official stand is that libvirt-python must always be >= libvirt.so.  Given that and the versions we have:

 * Ubuntu 3.6.0
 * Centos 3.9.0
 * Gentoo  ???
 * SuSE[1] ???


If we can get buy in from Dirk and Matthew that this u-c bump wont break those distros, or find a way to be compatible. then I think we're good to go.  My +1 can be treated as a +2 once we have positive reviews from Dirk and Matthew

[1] I know SuSE used to be the correct capitalisation is it still?  @Dirk?

Patch-set: 3
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1