Proposing distro for global-requirements
In the spirit of "python 3" first I would like to propose distro for
global-requirements.
The builtin platform.linux_distribution[1] is deprecated and will be
removed in 3.8 and the recommended replacement is distro. Since
linux_distribution is widely used in OpenStack projects[2], we should
have an approved alternative.

Is the library actively maintained?
Yes[3]

Is the library good code?
Yes, it is pretty simple code with 98% coverage[4]

Is the library license compatible?
Yes, Apache 2.0[5]

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?
Ubuntu since artful (17.10)
Fedora since 23
RHEL 6

Is the function of this library already covered by other libraries in
global-requirements.txt?
Not that I am aware of.

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?
As noted above[2], many OpenStack projects are using the deprecated
function. Octavia is one that would like to move to distro.

[1] https://docs.python.org/2/library/platform.html \
    #platform.linux_distribution
[2] http://codesearch.openstack.org/?q=platform.linux_distribution
[3] https://pypi.org/project/distro/1.2.0/#history
[4] https://codecov.io/github/nir0s/distro?branch=master
[5] https://github.com/nir0s/distro/blob/master/LICENSE

Change-Id: Iee61fad7954a56b225221e5b5254ee8094911d43