Add openshift to global-requirements
Is the library actively maintained?

Is the library good code?

Is the library python 3 compatible?

Is the library license compatible?
    Yes - Apache 2.0

Is the library already packaged in the distros we target (Ubuntu latest
    / Fedora latest)?

Is the function of this library already covered by other libraries in
    global-requirements.txt?
Is the library required for OpenStack project or related dev or
    infrastructure setup? (Answer to this should be Yes, of course) Which?

Change-Id: I36e2d1e78636964bcbd86af78a24b2bb623f202a
Co-Authored-By: Yossi Boaron <yossi.boaron.dev@gmail.com>