Add openshift to global-requirements
The kuryr-tempest-plugin uses openshift python client to verify
the functionality of Openshift resources.

Is the library actively maintained?
Yes - https://github.com/openshift/openshift-restclient-python
They appear to have made regular releases since April 2017 averaging
~1 per month.

Is the library good code?
Yes - and the documentation appears to cover all Openshift resources.

Is the library python 3 compatible?
Yes

Is the library license compatible?
Yes - Apache 2.0

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?
As far as I know, the answer is No

Is the function of this library already covered by other libraries in
global-requirements.txt?
No, the Openshift python client is the only package that can be used
to provision Openshift resources.

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes - kuryr-tempest-plugin uses openshift python client to verify the
functionality of Openshift resources.

Change-Id: I36e2d1e78636964bcbd86af78a24b2bb623f202a
Co-Authored-By: Yossi Boaron <yossi.boaron.dev@gmail.com>