Update patch set 6
Patch Set 6: Code-Review-1

So that came from: I99ad9f0b1d6c150bee6d6d35f84a74e99ab849b8 which pretty much makes sense but doesn't help us see *why* it's in lower-constraints.txt and not in upper-constraints.txt.

I'm currently building a lower constraint set *without* ruamel.yaml  If it ends up being installed we'll know that it's valid and what needs it.  Last time that took a couple of hours to build so I guess I'll post the results tomorrow.

I'm going to leave a -1 with this comment just because the problem needs to be understood so we can do a risk assessment.

Patch-set: 6