[DNM] Use docker images to generate constraints for multiple python versions
Currently when we generate constraints we create a venv for each
(system) python supllied on the commandline, install all of
global-requirements into that venv and capture the pip freeze.

Where this falls down is if we want to generate a freeze for python 3.4
and 3.5 we need an image that has both of those.  We cheated and just
'clone' them so if python3 is 3.4 we copy the results to 3.5 and vice
versa.  This kinda worked for a while but it has drawbacks.

This change provides a tool to generate docker images for pythons will
all of g-r installed.  It's a very rough PoC to show how things could
work.  There are lots of open questions:

- How do we make this nodepool/cloud provider friendly ?
- Can we run a registry per provider?
- Can we generate and caches these images and only run pip install -U
  g-r to speed up the build
- Are we okay with using docker this way?

I think it's clear we need to do something but what?

Change-Id: I751209a27f650b1c5849d6ac17bf5b322f02e40a