Update patch set 5
Patch Set 5:

> If we're not sure why we're preventing them, we should wait and see
 > if they actually cause problems.
 > 
 > I know we have avoided supporting pre-releases or alphas & betas in
 > the past because you have to give pip special instructions to
 > install them. I don't think that applies to post versioned
 > packages, since those are only supposed to have minor fixes for the
 > package itself.

Passing on the discussion on *all* .post releases which I think we need to have.

pyldap-3.0.0 is an *empty* package that just requires py-ldap-$some_version.  The only differences between 3.0.0 and 3.0.0.post1 are documentation and packaging.

 * afcd69d (HEAD -> master, tag: python-ldap-3.0.0.post1, origin/master, origin/HEAD) Bump version to 3.0.0.post1 to accomodate README/description change
 * 8e2aeee Add a warning about pip upgrade to README and package description
 * 47db63d Update python-ldap version and links


 README   | 28 ++++++++++++++++++++++++++--
 setup.py | 10 +++++-----
 2 files changed, 31 insertions(+), 7 deletions(-)

I think in the case of pyldap we really don't care about the .post1 as it's functionally identical to 3.0.0 and we should accept it in the next bot update.

I think we should think about a general policy and discuss it at the next team meeting.

Patch-set: 5
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0