Update patch set 2
Patch Set 2: Workflow-1

The fix is going to be OK for py36 but will break again when we start doing py37 work. Not sure if we want to preempt that issue, or wait for it to fail again?

Patch-set: 2
Reviewer: Gerrit User 8099 <8099@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1