Remove PyKafka from global-requirements
PyKafka was added to g-r as planned replacement for kafka-python client
in Monasca [1]. Support for this client was never added though and the
library is not used. Recent tests have shown that confluent-kafka client
provides better throughtput and Kafka API support. Current plan is to
add support for confluent-kafka client in Monasca instead of PyKafka
[2].

No other project uses PyKafka [3].

[1] https://git.openstack.org/cgit/openstack/requirements/commit/?id=c1749152a85a80180198a5985d34d3ab86aeb722
[2] https://storyboard.openstack.org/#!/story/2003079
[3] http://codesearch.openstack.org/?q=pykafka&i=nope&files=&repos=

Change-Id: I91202fc1187c284c0fd994ad7d8bd649f5973196