Update patch set 1
Patch Set 1:

Matthew, as I have written in PyKafka removal review [1] this library isn't used anywhere.
The work we start now won't be released in Rocky and is planned for the next cycle. I'm fine with waiting until the stable branch on requirements repo is created and FF lifted.

[1] https://review.openstack.org/585398

Patch-set: 1
Label: Verified=0