Update patch set 1
Patch Set 1: Code-Review+2

configparser is part of the standard library, and was renamed between Python 2 and Python 3. We recently updated all of the jobs to run under python 3 in https://review.openstack.org/#/c/578943/. The pip-install environment is affected because that patch changes the testenv default settings, and we didn't catch the problem in the gate because we do not run a job that uses that virtualenv in the gate.

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2