Update patch set 1
Patch Set 1: Code-Review+2

> > ok, so this was blocking the infra job then?
 > 
 > I don't think we use that environment for any jobs in CI. I'm not
 > sure what it's for.

It was added in Ibc37593afcc4d9f820cb88168e1aa15e773b2087 with the intent of making a voting job that called it.  IT was in response to several of our tools using libraries that weren't listed in requirements.txt (but were in test-requirements.txt).

I know that we're frozen but IMO that really is with the aim of stabilising the OpenStack release.  this change has no impact on that goal, and fixes a clear, albeit small, issue.

Patch-set: 1
Label: Code-Review=+2