Update patch set 1
Patch Set 1:

> do we still intend this to be run on py27?  does this code change
 > run correctly on py27?

Not really.  We can pick which python version we want it to run under.  I s'pose we could introduce incompatibilities and/or fail to notice a regression by only running it under a single python version but I'm okay with accepting that risk.  Esp. given we don't actually run it in the gate anyway.

Patch-set: 1