Update patch set 1
Patch Set 1:

> Beginning Rocky, Octavia offers a driver model for LB vendors. Any
 > vendor who would like to integrate a driver from his own repo will
 > need this dependency - VMware happened to be the 1st.

Octavia is being published to PyPI, so I see no issue with adding it to the dependency list to make it easier for third-party testing.

The current validation error is happening because this patch also needs to include a version for Octavia in the upper-constraints.txt file. Today the highest version that could be used is 2.0.1, but based on your comment above I doubt that is what anyone actually wants. I suggest waiting until after the final Rocky release is completed and then using 3.0.0.

I also think the Octavia team should consider switching the release model from cycle-with-milestones to cycle-with-intermediary so they can publish updates more often.

And of course we shouldn't approve the change until after we branch the requirements list for Rocky.

Patch-set: 1