Update patch set 1
Patch Set 1:

Why would we switch release models? We have no driver to release more often or drop from the milestone releases. I think we would only consider that if neutron, nova, and glance do so.

I don't have a problem with adding it, as I see neutron is included in the GR list.
That said, I'm not sure I understand the need for this. The driver package should be able to include Octavia in the requirements.txt of the driver package.

Patch-set: 1