Update patch set 1
Patch Set 1:

ya, was just going to mention the release model thing, it's generally a rule that libraries (which it sounds like octavia has become...) need that for the reason doug mentioned (releases during the cycle, beta, rc) cannot be installed by normal methods.  This would interfere with development of stuff dependant on octavia.

Again, I highly recommend that anyone wishing to use a service as a library push for the creation of $service-lib.  That way the service and the plugin can just depend on an actual library instead of having something try to provide both functions.

Patch-set: 1