Update patch set 1
Patch Set 1:

This is not how we handle drivers for neutron today, so why would octavia be different?
They install via setuptools entry point and have a defined interface.
I don't see why this requires a change in release cadence unless we do that to all of the OpenStack services.
They are not importing our code aside from a callback that is an interim solution until we decide on the security approach for the driver endpoint, at which point that goes away too.

Patch-set: 1