Update patch set 1
Patch Set 1:

> Absolutely NOT. Octavia is not a library and does not support or
 > encourage that.
 > Who said it was?

The only reason to include Octavia here is to allow a project to express a dependency on it. If a driver imports Octavia code, that makes that part of Octavia a library. If the Octavia team doesn't want to extract that into its own library, then the releases for the service need to be treated as libraries (i.e., when a new feature is available, prepare a release so drivers can depend on that new version). So Octavia itself isn't a library, but we need to treat it as one for testing purposes.

Patch-set: 1