Update patch set 1
Patch Set 1: Code-Review-1

Yeah, Octavia is not a library and has no intentions of becoming one.

I am -1 because this project is not a library and should not be treated as one. There should not be any confusion over that.

Patch-set: 1
Reviewer: Gerrit User 11628 <11628@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1