Bump eventlet to 0.24.1
We want this patch [0] in eventlet that avoids contacting nameservers
if there's at least one (IPv4 or IPv6) entry in the hosts file.
In the case where no nameservers are configured, this will avoid
unnecessary delays when resolving hostnames.

Also eventlet <v0.21.0 is affected by [1] on Python 3.6 causing
neutron to fail with SSL enabled. The issue is fixed in v0.21.0

Also updates os-xenapi to compatible version.

[0] https://github.com/eventlet/eventlet/commit/10ff67f84a1e2f53c8eb254944ebf81762041b45
[1] https://github.com/eventlet/eventlet/issues/371

Closes-bug: #1785615
Closes-bug: #1884808
Change-Id: Ia18372625d1e5c87a3afdb5617f0390c19d46470