Modified lower constraints for openstacksdk
Updated openstacksdk lower constraints from 0.9.19 to 0.11.0
because python-masakariclient and masakari-monitors fails to
create connection as service descriptors are not supported in
these versions and profile parameter doesn't work at all.

This is a blocker issue [1] in masakariclient and
masakari-monitors for stable/queens branch. A patch is proposed
to fix this issue [2] and it works with openstacksdk
versions >= 0.11.0.

[1]: https://bugs.launchpad.net/masakari/+bug/1779752
[2]: https://review.openstack.org/#/c/588157

Change-Id: I24ec493f3333a5b3f6bfed9395c921fb28500065