Update patch set 1
Patch Set 1:

Hi Matthew,

openstacksdk is widely used by many OpenStack services but we need this fix badly for masakari which is a instance-ha service.

For masakari service, openstacksdk is used in python-masakariclient [1] and masakari-monitors [2].

Note: In masakari-monitors even though openstacksdk is used, it wasn't mentioned in global requirements i.e. requirements.txt in stable/queens branch. This issue has been fixed in the Rocky.

[1] : https://github.com/openstack/python-masakariclient/blob/stable/queens/requirements.txt#L5

[2] : https://github.com/openstack/masakari-monitors/blob/stable/queens/masakarimonitors/ha/masakari.py#L16

Patch-set: 1