Update patch set 1
Patch Set 1:

Hi Matthew,

Unfortunately python-masakariclient is not using osc_lib and all sdk code  was part of python-masakariclient [2] itself in stable/queens. python-masakariclient and masakari-monitors are completely broken in stable/queens due to proxy related changes done in openstacksdk [3] in Queens release. These changes done in SDK weren't addressed properly in python-masakariclient and masakari-monitors. Now, we are fixing this issue in stable/queens in patch [1].

Patch [1] fixes this issue in python-masakariclient for sdk version >=0.11.0 but in 0.9.19 and 0.10.0 sdk versions, it's failing to create connection object as explained in the commit message.

[1] : https://review.openstack.org/#/c/588157
[2] : https://github.com/openstack/python-masakariclient/tree/stable/queens/masakariclient/sdk
[3] : https://github.com/openstack/openstacksdk/commit/f9b96861577e26f0540158e00706e2505213f4bf#diff-0fbc1bc8e7b00464f7eb4049a015479a

Patch-set: 1