add pydot dependency
Is the library actively maintained?

   It appears that pydot has been updated more recently than pydot3,
   although neither seem particularly current.

Is the library good code?

   No idea.

Is the library python3-compatible?

   It claims support through python 3.7.

Is the library license compatible?

   Yes, MIT

Is the library already packaged in the distros we target?

   Unknown. We've apparently been using it in taskflow for a while now,
   though.

Is the function of this library already covered by other libraries in
global-requirements.txt?

   We already have pydot3, but the version of networkx we are using does
   not like the version number of pydot3. We worked around that by
   placing the dependency in tox.ini, where the requirements check job
   didn't see it. This patch is meant to let us start managing the
   dependency properly.

Is the library required for OpenStack project or related dev or
infrastructure setup?

   Yes, taskflow uses it.

If the library release is managed by the Openstack release process does
it use the cycle-with-intermediary release type?

   This is not an OpenStack library.

Change-Id: I18bd5819b55a574eefd57c6550c4d09d422ac4f6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>