Add Django-debreach to global-requirements.txt
This provides Basic/extra mitigation against the BREACH attack for
Django projects.

More details- https://github.com/lpomfrey/django-debreach

The blueprint has been approved to add this-
https://blueprints.launchpad.net/horizon/+spec/mitigate-breach-attacks

It is being targetted for stein.

Is the library actively maintained?

Last release Dec 21, 2017 and no indications of project being abandoned.

Is the library good code?

We at AT&T have been using this code from quite sometime and have faced
no issues.

Is the library license compatible?

Yes.

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

Yes.

Is the function of this library already covered by other libraries in global-requirements.txt?

No.

Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?

Yes, this is needed for Projects based on Django.

If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?

No.

Change-Id: I32f11e089fc794444ef267b463c7fb2ad8cfa96a