Add sphinxcontrib-programoutput
This sphinx extension provides a way to document
CLI invocations. For example, see the python-manilaclient
change here [1]

The upstream code repository is maintained on GitHub, and
has been quite stable. It has continuous integration [3]
and accepts issues and PRs. It is tested with python 2.7 and
python3. The 0.11 (latest release) has been packaged in Ubuntu,
Fedora, Gentoo and SUSE.

There are several OpenStack projects using this in their
documentation, please see [4].

[1] https://review.openstack.org/#/c/595362/
[2] https://github.com/NextThought/sphinxcontrib-programoutput
[3] https://travis-ci.org/NextThought/sphinxcontrib-programoutput
[4] http://codesearch.openstack.org/?q=programoutput&i=nope&files=&repos=

Change-Id: I80b9d0b0f027bbffc9324030329dad3ad0a2ca5a