Update patch set 1
Patch Set 1:

In the past this file managed the projects that would recieve updates from the bot when global-requirements changed.  During the rocky cycle we disabled that.

It looks to be that the only thing that consumes projects.txt is for updates from openstack-manuals again that job is deprecated.

So while I have no problem adding your project to this list, I'd like to be sure that we're doign it for the right reasons.  It maybe we have out dated documentation that ask you to do this, which we should fix.

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0