Update patch set 3
Patch Set 3:

Okay so the hack sort of works but it isn't enough.  We install ceilometer[1] and networking-odl[2] but then when we try to generate the constraints for both python 2 and 3 we again fail to find ceilometer on pypi[3].

We might be able to disable the constraints generation as the have the check-uc-* jobs also running and it's some level a duplication of effort.  If we did that it'd need a massive comment.


[1] http://logs.openstack.org/77/599277/3/check/requirements-integration/8ef07ce/job-output.txt.gz#_2018-09-03_22_04_29_788704
[2] http://logs.openstack.org/77/599277/3/check/requirements-integration/8ef07ce/job-output.txt.gz#_2018-09-03_22_06_46_667090
[3] http://logs.openstack.org/77/599277/3/check/requirements-integration/8ef07ce/job-output.txt.gz#_2018-09-03_22_14_23_502018

Patch-set: 3
Label: Verified=0